save before the storm

+++ cpp06/ex00/ScalarConverter.hpp
@@ -6,7 +6,7 @@
 /*   By: yloutfi <yloutfi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/02 11:16:13 by yloutfi           #+#    #+#             */
-/*   Updated: 2023/12/05 11:48:13 by yloutfi          ###   ########.fr       */
+/*   Updated: 2023/12/06 11:18:01 by yloutfi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,10 @@
 #define __SCALCONVERT__
 
 #include <iostream>
+#include <iomanip>
+#include <climits>
 #include <cstring>
+#include <cstdlib>
 
 class ScalarConverter
 {
@@ -28,7 +31,7 @@ class ScalarConverter
 
 		static void convert(std::string arg);
 		static void toChar(char c);
-		static void toInt(int i);
+		static void toInt(float i);
 		static void toDouble(double d);
 		static void toFloat(float f);
 };
